#67 refactor : 스웨거 분류 작업
