+++ src/main/java/com/seoulmilk/seoulmilkServer/domain/agency/controller/AgencyController.java
@@ -1,0 +1,112 @@
+package com.seoulmilk.seoulmilkServer.domain.agency.controller;
+
+import com.seoulmilk.seoulmilkServer.domain.agency.domain.Agency;
+import com.seoulmilk.seoulmilkServer.domain.agency.service.AgencyAuthService;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.domain.NtsTax;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.domain.enums.IsSuccess;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.request.SubmitNtxTaxRequestDTO;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.request.UpdateNtsTaxRequestDTO;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.response.GetNtsTaxListResponseDTO;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.response.GetOcrNtsTaxListResponseDTO;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.response.UpdateNtsTaxResponseDTO;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.service.NtsTaxCommandService;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.service.NtsTaxQueryService;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.service.NtxTaxMappingService;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.service.OcrService;
+import com.seoulmilk.seoulmilkServer.global.common.ApiResponse;
+import com.seoulmilk.seoulmilkServer.global.error.ErrorCode;
+import com.seoulmilk.seoulmilkServer.global.error.exception.BusinessException;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import java.time.LocalDate;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/agency")
+@Tag(name = "[대리점 API]")
+public class AgencyController {
+
+    private final OcrService ocrService;
+    private final AgencyAuthService agencyAuthService;
+    private final NtsTaxCommandService ntsTaxCommandService;
+    private final NtsTaxQueryService ntsTaxQueryService;
+    private final NtxTaxMappingService ntxTaxMappingService;
+
+
+    @Operation(summary = "대리점 - 세금 계산서 OCR")
+    @PostMapping(value = "/nts-tax/upload", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    public ApiResponse<GetOcrNtsTaxListResponseDTO> getOcrTest(
+        @RequestParam("files") List<MultipartFile> files) {
+
+        if (files.isEmpty()) {
+            throw new BusinessException(ErrorCode.NTS_TAX_NOT_UPLOAD);
+        }
+
+        return ApiResponse.success(ocrService.getOcrResponse(files));
+    }
+
+    @Operation(summary = "대리점 - 세금 계산서 수정")
+    @PutMapping("/nts-tax/edit")
+    public ApiResponse<UpdateNtsTaxResponseDTO> updateNtsTax(
+        @RequestBody UpdateNtsTaxRequestDTO request) {
+
+        Agency agency = agencyAuthService.getCurrentAgency();
+
+        return ApiResponse.success(ntsTaxCommandService.updateNtsTax(agency, request));
+    }
+
+    @Operation(summary = "대리점 - 세금 계산서 목록 조회")
+    @GetMapping("/nts-tax")
+    public ApiResponse<GetNtsTaxListResponseDTO.NtsTaxListResponseDTO> getNtsTaxList(
+        @RequestParam(name = "page") Integer page,
+        @RequestParam(name = "isSuccess") IsSuccess isSuccess) {
+        Agency agency = agencyAuthService.getCurrentAgency();
+
+        return ApiResponse.success(ntsTaxQueryService.getNtsTaxList(agency, page, isSuccess));
+    }
+
+    @Operation(summary = "대리점 - 세금 계산서 검색")
+    @GetMapping("/nts-tax/search")
+    public ApiResponse<GetNtsTaxListResponseDTO.SearchNtsTaxListResponseDTO> searchNtsTaxList(
+        @RequestParam(name = "page") Integer page,
+        @RequestParam(required = false) LocalDate startDate,
+        @RequestParam(required = false) LocalDate endDate,
+        @RequestParam(required = false) List<String> ipNameList) {
+        Agency agency = agencyAuthService.getCurrentAgency();
+
+        Page<NtsTax> ntsTaxList = ntsTaxQueryService.searchNtsTaxList(agency, page, startDate,
+            endDate, ipNameList);
+
+        return ApiResponse.success(GetNtsTaxListResponseDTO.from(ntsTaxList));
+    }
+
+    @Operation(summary = "대리점 - 선택한 세금 계산서 제출")
+     @PostMapping("/nts-tax/submit-hometax")
+     public ResponseEntity submitNtsTaxList(@RequestBody SubmitNtxTaxRequestDTO request) {
+         ntxTaxMappingService.submitNtxTax(request);
+         return ResponseEntity.ok().body("세금계산서 제출 완료");
+     }
+
+
+     @Operation(summary = "대리점 - 전체 세금 계산서 제출")
+     @PostMapping("/nts-tax/submit-hometax/all")
+     public ResponseEntity submitAllNtsTax()
+     {
+         ntxTaxMappingService.submitAllNtsTax();
+         return ResponseEntity.ok().body("전체 세금계산서 제출 완료");
+     }
+
+}
