+++ src/main/java/com/seoulmilk/seoulmilkServer/domain/ntsTax/controller/NtsTaxController.java
@@ -4,44 +4,22 @@ import com.seoulmilk.seoulmilkServer.domain.agency.domain.Agency;
 import com.seoulmilk.seoulmilkServer.domain.agency.service.AgencyAuthService;
 import com.seoulmilk.seoulmilkServer.domain.member.domain.Member;
 import com.seoulmilk.seoulmilkServer.domain.member.service.MemberAuthService;
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.domain.NtsTax;
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.domain.enums.IsSuccess;
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.domain.enums.Status;
 import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.request.DeleteNtsTaxRequestDTO;
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.request.SubmitNtxTaxRequestDTO;
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.request.UpdateNtsTaxRequestDTO;
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.response.GetHometaxResponseDTO;
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.response.GetNtsTaxListResponseDTO;
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.response.GetOcrNtsTaxListResponseDTO;
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.response.UpdateNtsTaxResponseDTO;
 import com.seoulmilk.seoulmilkServer.domain.ntsTax.service.NtsTaxCommandService;
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.service.NtsTaxQueryService;
 import com.seoulmilk.seoulmilkServer.domain.ntsTax.service.NtxTaxMappingService;
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.service.OcrService;
 import com.seoulmilk.seoulmilkServer.global.common.ApiResponse;
-import com.seoulmilk.seoulmilkServer.global.error.ErrorCode;
-import com.seoulmilk.seoulmilkServer.global.error.exception.BusinessException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
 
-@Tag(name = "[세금 계산서 및 OCR]")
+@Tag(name = "[사원, 대리점 공통 API]")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api")
@@ -51,56 +29,7 @@ public class NtsTaxController {
     private final MemberAuthService memberAuthService;
     private final AgencyAuthService agencyAuthService;
     private final NtsTaxCommandService ntsTaxCommandService;
-    private final OcrService ocrService;
-    private final NtsTaxQueryService ntsTaxQueryService;
     private final NtxTaxMappingService ntxTaxMappingService;
-
-    @Operation(summary = "대리점 - 세금 계산서 OCR")
-    @PostMapping(value = "/agency/nts-tax/upload", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ApiResponse<GetOcrNtsTaxListResponseDTO> getOcrTest(
-        @RequestParam("files") List<MultipartFile> files) {
-
-        if (files.isEmpty()) {
-            throw new BusinessException(ErrorCode.NTS_TAX_NOT_UPLOAD);
-        }
-
-        return ApiResponse.success(ocrService.getOcrResponse(files));
-    }
-
-    @Operation(summary = "대리점 - 세금 계산서 수정")
-    @PutMapping("/agency/nts-tax/edit")
-    public ApiResponse<UpdateNtsTaxResponseDTO> updateNtsTax(
-        @RequestBody UpdateNtsTaxRequestDTO request) {
-
-        Agency agency = agencyAuthService.getCurrentAgency();
-
-        return ApiResponse.success(ntsTaxCommandService.updateNtsTax(agency, request));
-    }
-
-    @Operation(summary = "대리점 - 세금 계산서 목록 조회")
-    @GetMapping("/agency/nts-tax")
-    public ApiResponse<GetNtsTaxListResponseDTO.NtsTaxListResponseDTO> getNtsTaxList(
-        @RequestParam(name = "page") Integer page,
-        @RequestParam(name = "isSuccess") IsSuccess isSuccess) {
-        Agency agency = agencyAuthService.getCurrentAgency();
-
-        return ApiResponse.success(ntsTaxQueryService.getNtsTaxList(agency, page, isSuccess));
-    }
-
-    @Operation(summary = "대리점 - 세금 계산서 검색")
-    @GetMapping("/agency/nts-tax/search")
-    public ApiResponse<GetNtsTaxListResponseDTO.SearchNtsTaxListResponseDTO> searchNtsTaxList(
-        @RequestParam(name = "page") Integer page,
-        @RequestParam(required = false) LocalDate startDate,
-        @RequestParam(required = false) LocalDate endDate,
-        @RequestParam(required = false) List<String> ipNameList) {
-        Agency agency = agencyAuthService.getCurrentAgency();
-
-        Page<NtsTax> ntsTaxList = ntsTaxQueryService.searchNtsTaxList(agency, page, startDate,
-            endDate, ipNameList);
-
-        return ApiResponse.success(GetNtsTaxListResponseDTO.from(ntsTaxList));
-    }
 
     @Operation(summary = "본사, 대리점 - 세금 계산서 단건 삭제")
     @DeleteMapping("/nts-tax/{nts_tax_id}")
@@ -122,59 +51,6 @@ public class NtsTaxController {
 
         return ApiResponse.success("NtsTaxList Deletion successful");
     }
-
-    @Operation(summary = "대리점 - 선택한 세금 계산서 제출")
-    @PostMapping("/agency/nts-tax/submit-hometax")
-    public ResponseEntity submitNtsTaxList(@RequestBody SubmitNtxTaxRequestDTO request) {
-        ntxTaxMappingService.submitNtxTax(request);
-        return ResponseEntity.ok().body("세금계산서 제출 완료");
-    }
-
-
-    @Operation(summary = "대리점 - 전체 세금 계산서 제출")
-    @PostMapping("/agency/nts-tax/submit-hometax/all")
-    public ResponseEntity submitAllNtsTax()
-    {
-        ntxTaxMappingService.submitAllNtsTax();
-        return ResponseEntity.ok().body("전체 세금계산서 제출 완료");
-    }
-
-    @Operation(summary = "본사 - 세금 계산서 이번 달 내역 조회")
-    @GetMapping("/employee/nts-tax/view-hometax/recent")
-    public ApiResponse<GetHometaxResponseDTO.GetHometaxListResponseDTO> getHometaxList(
-        @RequestParam(name = "page") Integer page,
-        @RequestParam(name = "isSuccess") Status status) {
-        Member member = memberAuthService.getCurrentMember();
-
-        return ApiResponse.success(ntsTaxQueryService.getHometaxList(member, page, status));
-    }
-
-    @Operation(summary = "본사 - 세금 계산서 전체 내역 통합 조회")
-    @GetMapping("/employee/nts-tax/view-hometax/history")
-    public ApiResponse<GetHometaxResponseDTO.GetHometaxListResponseDTO> getHometaxHistory(
-        @RequestParam(name = "page") Integer page,
-        @RequestParam(required = false) Status status) {
-        Member member = memberAuthService.getCurrentMember();
-
-        return ApiResponse.success(ntsTaxQueryService.getHometaxHistory(member, page, status));
-    }
-
-    @Operation(summary = "본사 - 세금 계산서 내역 검색")
-    @GetMapping("/employee/employee/nts-tax/search-hometax")
-    public ApiResponse<GetHometaxResponseDTO.GetHometaxListResponseDTO> searchHometaxList(
-        @RequestParam(name = "page") Integer page,
-        @RequestParam(required = false) LocalDate startMonth,
-        @RequestParam(required = false) LocalDate endMonth,
-        @RequestParam(required = false) List<String> suNameList,
-        @RequestParam(required = false) List<String> ipNameList) {
-        Member member = memberAuthService.getCurrentMember();
-
-        Page<NtsTax> searchHometaxList = ntsTaxQueryService.searchHometaxList(member, page,
-            startMonth, endMonth, suNameList, ipNameList);
-
-        return ApiResponse.success(GetHometaxResponseDTO.from(searchHometaxList));
-    }
-
 
     //    @Operation(summary = "세금계산서 홈택스 진위 여부 확인 [단건]")
     //    @PostMapping("/nts-tax/hometax")
